# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: LearnFlow API
  version: 1.0.0
  description: AI-powered learning platform API
servers:
  - url: http://localhost:3000
paths:
...
  /api/v1/courses:
    get:
      summary: List courses
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Course list
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
    post:
      summary: Create/generate course
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [topic]
              properties:
                topic: { type: string }
                title: { type: string }
                description: { type: string }
                depth: { type: string }
                max_lessons: { type: number }
      responses:
        '201':
          description: Course created
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '402':
          description: Payment required (free tier limit)
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/courses/{id}:
    get:
      summary: Get course detail
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Course detail
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '404':
          description: Not found
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      summary: Delete course
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Deleted
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '404':
          description: Not found
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /api/v1/marketplace/courses:
    get:
      summary: Browse course marketplace
      parameters:
        - name: keyword
          in: query
          schema: { type: string }
        - name: topic
          in: query
          schema: { type: string }
        - name: difficulty
          in: query
          schema: { type: string }
        - name: maxPrice
          in: query
          schema: { type: number }
      responses:
        '200':
          description: Marketplace courses
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
  /api/v1/marketplace/courses/{id}:
    get:
      summary: Marketplace course detail
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Course
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '404':
          description: Not found
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/marketplace/publish:
    post:
      summary: Publish a course
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                [title, topic, description, difficulty, price, lessonCount, attributionCount]
              properties:
                title: { type: string }
                topic: { type: string }
                description: { type: string }
                difficulty: { type: string }
                price: { type: number }
                lessonCount: { type: number }
                attributionCount: { type: number }
                readabilityScore: { type: number }
      responses:
        '201':
          description: Published
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
  /api/v1/marketplace/courses/{id}/reviews:
    post:
      summary: Create a course review
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rating]
              properties:
                rating: { type: integer, minimum: 1, maximum: 5 }
                text: { type: string }
      responses:
        '201':
          description: Created
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
  /api/v1/marketplace/checkout:
    post:
      summary: Start checkout (mock)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [courseId]
              properties:
                courseId: { type: string }
      responses:
        '200':
          description: Created
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
  /api/v1/marketplace/checkout/confirm:
    post:
      summary: Confirm checkout (mock)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [paymentIntentId]
              properties:
                paymentIntentId: { type: string }
      responses:
        '200':
          description: Confirmed
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
...
  /api/v1/marketplace/agents/submit:
    post:
      summary: Submit an agent
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                manifest: { type: object }
      responses:
        '201':
          description: Submitted
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
  /api/v1/marketplace/agents/activated:
    get:
      summary: List activated agents for current user
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Activated agent IDs
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
  /api/v1/marketplace/agents/{id}/activate:
    post:
      summary: Activate marketplace agent
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Agent activated
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
  /api/v1/marketplace/creator/dashboard:
    get:
      summary: Creator dashboard
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Dashboard
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    XRequestId:
      description: Request identifier for tracing
      schema:
        type: string
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error, requestId]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
        requestId:
          type: string
        status:
          type: integer
...